Separate test code by type
- unit tests
- integration tests
