+++ src/test/unit/path.test.ts
@@ -1,5 +1,5 @@
 import * as assert from 'assert';
-import * as path from '../path';
+import * as path from '../../path';
 
 
 suite("path", () => {
